fix: update test files to use Jest global helpers
- Add setupTestDom and resetTestMocks to all failing test files
- Split large test files into smaller modules to prevent worker termination
- Update beforeEach blocks to properly set up test environment
- Improve test structure for better reliability in CI
