+++ .gitignore
@@ -58,3 +58,5 @@ logs
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+
+**/.claude/settings.local.json
