--- T001-plan.md
@@ -1,40 +1,0 @@
-# T001 Plan: Sanitize User Inputs Before Validation and Saving
-
-## Overview
-Implement input sanitization for all user inputs in the options form to prevent XSS attacks and other injection vulnerabilities.
-
-## Current State Analysis
-Let's examine the current implementation of `formHandler.js` to understand:
-1. Which user inputs need sanitization
-2. The current validation logic
-3. Where to insert sanitization (before validation)
-
-## Implementation Plan
-
-### 1. Create Sanitization Functions
-- Implement a general-purpose `sanitizeInput` function to handle common sanitization tasks
-- Create specific sanitization functions for each input type (if needed)
-- Use appropriate sanitization techniques:
-  - HTML entity escaping for text inputs
-  - Strict whitelisting for numeric inputs and special characters
-
-### 2. Update Form Handler Logic
-- Apply sanitization to all user inputs before validation:
-  - `currencySymbol`: Sanitize to prevent XSS in symbol display
-  - `amount`: Ensure only valid numeric characters
-  - `debounceInterval`: Ensure only valid numeric characters
-  - Any other user-provided inputs
-
-### 3. Add Tests
-- Create test cases to verify that malicious inputs are properly sanitized
-- Attempt common XSS payloads and ensure they're neutralized
-
-## Approach Details
-1. For currency symbols and codes: Use a whitelist approach that only allows valid characters
-2. For numeric inputs: Ensure inputs are numeric and within reasonable bounds
-3. Apply sanitization early in the input processing flow, before validation and storage
-
-## Testing Strategy
-- Unit tests for each sanitization function with various inputs
-- Integration tests that verify the full form handling process with sanitization
-- Specific tests for known XSS vectors and edge cases
