--- T016-plan.md
@@ -1,38 +1,0 @@
-# T016 Plan - Add "type": "module" to package.json
-
-## Task Details
-- **ID:** T016
-- **Type:** Refactor
-- **Priority:** P1
-- **Context:** cr-16, Step 1
-- **Status:** In Progress
-
-## Current State Analysis
-The codebase is transitioning to ES6 modules but currently doesn't have `"type": "module"` set in package.json. This means Node.js treats .js files as CommonJS modules by default.
-
-## Plan
-1. Review the current package.json file
-2. Add `"type": "module"` to the root level of package.json
-3. Verify the change doesn't break existing scripts or builds
-4. Commit the change
-
-## Considerations
-- This change will impact how Node.js interprets .js files (as ES6 modules rather than CommonJS by default)
-- It's the first step in a multi-step process to fully adopt ES6 modules across the codebase
-- Subsequent tasks (T017, T018) will update build configs and tests to use ES6 module syntax
-
-## Risks
-- Potential to break scripts that rely on CommonJS module semantics
-- Build scripts and testing infrastructure may need additional updates
-
-## Implementation
-1. Add the `"type": "module"` property to package.json
-2. Run key npm scripts to verify functionality
-
-## Testing
-- Run `npm run build` to verify the build process still works
-- Basic sanity check of package.json syntax
-
-## Completion Criteria
-- package.json contains `"type": "module"`
-- Node treats JS files as ES6 modules by default
